Rename jwt modules
In addition, I did a small fix regarding debug logs when authorization fails.
Previously I logged the error when validation failed, but not when the actual extraction failed.
