+++ tower-oauth2-resource-server/src/jwt_validate.rs
@@ -4,7 +4,6 @@ use std::{
 };
 
 use async_trait::async_trait;
-use http::HeaderMap;
 use jsonwebtoken::{
     decode, decode_header,
     jwk::{AlgorithmParameters, Jwk, JwkSet, KeyAlgorithm},
@@ -19,25 +18,6 @@ use crate::{
     jwks::JwksConsumer,
     validation::ClaimsValidationSpec,
 };
-
-pub trait JwtExtractor {
-    fn extract_jwt(&self, headers: &HeaderMap) -> Result<String, AuthError>;
-}
-
-pub struct BearerTokenJwtExtractor;
-
-impl JwtExtractor for BearerTokenJwtExtractor {
-    fn extract_jwt(&self, headers: &HeaderMap) -> Result<String, AuthError> {
-        Ok(headers
-            .get(http::header::AUTHORIZATION)
-            .ok_or(AuthError::MissingAuthorizationHeader)?
-            .to_str()
-            .map_err(|_| AuthError::InvalidAuthorizationHeader)?
-            .strip_prefix("Bearer ")
-            .ok_or(AuthError::InvalidAuthorizationHeader)?
-            .to_owned())
-    }
-}
 
 #[async_trait]
 pub trait JwtValidator<Claims> {
